Refactor imports: move constants from core to shared directory for better organization
