+++ frontend/src/app/shared/components/footer/footer.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { NgOptimizedImage } from '@angular/common';
-import { RouteLink } from '../../../core/constants/route-link';
-import { LabelConstants } from '../../../core/constants/label.constants';
+import { RouteLink } from '../../../shared/constants/route-link';
+import { LabelConstants } from '../../../shared/constants/label.constants';
 
 @Component({
 	selector: 'app-footer',
