+++ frontend/src/app/shared/components/user-profile/user-profile.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { LabelConstants } from '../../../core/constants/label.constants';
+import { LabelConstants } from '../../../shared/constants/label.constants';
 import {
 	FormBuilder,
 	FormControl,
@@ -9,7 +9,7 @@ import {
 	Validators,
 } from '@angular/forms';
 import Swal from 'sweetalert2';
-import { ErrorMessageConstants } from '../../../core/constants/error-message.constants';
+import { ErrorMessageConstants } from '../../../shared/constants/error-message.constants';
 import { CommonModule } from '@angular/common';
 import { UserProfileFormFields } from './user-profile.interface';
 import { SideBarUserProfile } from '../side-bar-user-profile/side-bar-user-profile';
